+++ changelog.md
@@ -3,3 +3,5 @@
 - Added Search
 - Added Send and Get DM
 - Added Show Followers
+- Update Status without explicit response
+- Follow and Unfollow via Username